Extract publish/binary config and add to spider
Configure the publish and binary compatibility through the addOns
project to avoid duplication.
Add configuration to spider add-on.
Address JavaDoc errors in some classes.

Part of zaproxy/zaproxy#3113.
