+++ addOns/commonlib/gradle.properties
@@ -1,2 +1,8 @@
 version=1.11.0
 release=false
+
+zap.maven.publish=true
+zap.maven.pom.inceptionyear=2020
+
+zap.japicmp=true
+zap.japicmp.baseversion=1.0.0
