+++ addOns/fuzz/gradle.properties
@@ -1,2 +1,5 @@
 version=13.8.0
 release=false
+
+zap.maven.publish=true
+zap.maven.pom.inceptionyear=2015
