+++ addOns/network/gradle.properties
@@ -1,2 +1,5 @@
 version=0.3.0
 release=false
+
+zap.maven.publish=true
+zap.maven.pom.inceptionyear=2021
