+++ addOns/spider/gradle.properties
@@ -1,2 +1,5 @@
 version=0.1.0
 release=false
+
+zap.maven.publish=true
+zap.maven.pom.inceptionyear=2022
